Create high scores component; update game.ex logic to get data for component
