+++ project_quazar/lib/project_quazar_web/live/create_ship_live.ex
@@ -12,7 +12,7 @@ defmodule ProjectQuazarWeb.CreateShipLive do
   # It should return valid HEEx or LEEx markup.
   @impl true
   def render(assigns) do
-    ~L"""
+    ~H"""
     <div>
       <%= @sample_assign %>
     </div>
