+++ project_quazar/lib/project_quazar_web/live/game.ex
@@ -3,15 +3,24 @@ defmodule ProjectQuazarWeb.Game do
   use ProjectQuazarWeb, :live_view
   alias ProjectQuazarWeb.Presence
   alias ProjectQuazar.PubSub
+  alias ProjectQuazar.HighScores
 
   @presence "project_quazar:presence"
 
   @impl true
   def mount(_params, _session, socket) do
+
+    # For the all-time high score component, which subscribes to a PubSub
+    # which gets its data from an ETS table.
+    Phoenix.PubSub.subscribe(ProjectQuazar.PubSub, "high_scores:updates")
+    top_scores = fetch_top_scores()
+
     {:ok, socket
       |> assign(:joined, false)
       |> assign(:users, %{})
-      |> assign(:error_message, "")}
+      |> assign(:error_message, "")
+      |> assign(:top_scores, top_scores)
+    }
   end
 
   @impl true
@@ -38,6 +47,12 @@ defmodule ProjectQuazarWeb.Game do
       |> handle_joins(diff.joins)}
   end
 
+  @impl true
+  def handle_info(:scores_updated, socket) do
+    top_scores = fetch_top_scores()
+    {:noreply, assign(socket, :top_scores, top_scores)}
+  end
+
   defp handle_joins(socket, joins) do
     users = joins
       |> Enum.map(fn {user, %{metas: [meta | _]}} -> {user, meta} end)
@@ -58,4 +73,13 @@ defmodule ProjectQuazarWeb.Game do
       |> Enum.into(%{})
     assign(socket, :users, sorted_users)
   end
+
+  # Fetches top scores for all-time high scores component
+  defp fetch_top_scores do
+    case HighScores.fetch_top_scores() do
+      {:ok, top_scores} -> top_scores
+      {:error, _reason} -> []
+    end
+  end
+
 end
