+++ project_quazar/lib/project_quazar_web/live/game.html.heex
@@ -1,6 +1,6 @@
 <%# Master HTML Heex Page for all Front End %>
 <%# Main layout for the application %>
-<%# 
+<%#
   @joined: boolean - determines if the user has joined the game
     - true: renders the main game screen
     - false: renders the lobby screen
@@ -11,7 +11,7 @@
   <%# TODO: Add the game board here %>
 
   <%# Live scoreboard menu %>
-  <.live_component 
+  <.live_component
     module={ProjectQuazarWeb.Scoreboard}
     id="scoreboard"
     current_user={@current_user}
@@ -25,6 +25,11 @@
     module={ProjectQuazarWeb.UserForm}
     id="user-form"
     error_message={@error_message}
+  />
+  <.live_component
+    module={ProjectQuazarWeb.AllTimeHighScores}
+    id="all_time-high-scores-component"
+    top_scores={@top_scores}
   />
 
   <%# TODO: Add the all time high score screen here %>
