+++ project_quazar/lib/project_quazar_web/live/start_screen_live.ex
@@ -12,7 +12,7 @@ defmodule ProjectQuazarWeb.StartScreenLive do
   # It should return valid HEEx or LEEx markup.
   @impl true
   def render(assigns) do
-    ~L"""
+    ~H"""
     <div>
       <%= @sample_assign %>
     </div>
